Add DOCKER-USER chain when iptables=true is set
This PR fixes the regression introduced by
https://github.com/docker/libnetwork/pull/2339 to
correctly insert the DOCKER-USER chain if iptables=true
is set in the Daemon config

Addresses : https://github.com/docker/for-linux/issues/810
