--- firewall_others.go
@@ -1,6 +1,0 @@
-// +build !linux
-
-package libnetwork
-
-func (c *controller) arrangeUserFilterRule() {
-}
